Solucion error parametros en rotacionAntiohorario

+++ acciones_estructuras.h
@@ -7,6 +7,6 @@
 using namespace std;
 
 unsigned short** alinearmatriz(unsigned short n1, unsigned short n2, unsigned short rotaciones);
-void rotacionAntiHorario(int** matrix, int N, int rotaciones);
+void rotacionAntiHorario(unsigned short** matrix, int N, int rotaciones);
 
 #endif // ACCIONES_ESTRUCTURAS_H
